+++ native/asmchdr/gen_chdsect.sh
@@ -1,5 +1,4 @@
-#! #!/usr/bin/env bash
-set -e
+#!/usr/bin/env bash
 printf "\n"
 
 echo "Setting required env vars..."